backend/requirements.txt
Flask
Flask-SQLAlchemy
pandas
python-pptx
